[FIX] [Main] ktx 모듈 종속성 라이브러리 문제 해결 맟 issue7 해결 테스트
