+++ app/src/main/java/com/example/android_repo_04/viewmodel/LoginViewModel.kt
@@ -1,9 +1,7 @@
 package com.example.android_repo_04.viewmodel
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.*
+import androidx.lifecycle.viewmodel.CreationExtras
 import com.example.android_repo_04.BuildConfig
 import com.example.android_repo_04.api.GitHubLoginRepository
 import kotlinx.coroutines.CoroutineScope
@@ -16,10 +14,10 @@ class LoginViewModel(private val gitHubLoginRepository: GitHubLoginRepository) :
     val token: LiveData<String> get() = _token
 
     fun requestToken(code: String) {
-        CoroutineScope(Dispatchers.IO).launch {
+        viewModelScope.launch {
             gitHubLoginRepository.requestToken(BuildConfig.CLIENT_ID, BuildConfig.CLIENT_SECRET, code) { response ->
-                if (response.isSuccessful && response.body()?.accessToken != null) {
-                    _token.postValue(response.body()?.accessToken)
+                if (response.accessToken != "") {
+                    _token.postValue(response.accessToken)
                 } else {
                     _token.postValue("error")
                 }
@@ -29,7 +27,7 @@ class LoginViewModel(private val gitHubLoginRepository: GitHubLoginRepository) :
 }
 
 class LoginViewModelFactory(private val gitHubRepository: GitHubLoginRepository): ViewModelProvider.Factory {
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(LoginViewModel::class.java)) {
             return LoginViewModel(gitHubRepository) as T
         }
